mor reorg
