+++ app/commands/test_data.py
@@ -1,0 +1,73 @@
+import functools
+import uuid
+
+import click
+from flask import cli as flask_cli
+
+from app.dao.services_dao import (
+    dao_fetch_all_services_by_user,
+    delete_service_and_all_associated_db_objects,
+)
+from app.dao.users_dao import delete_model_user, delete_user_verify_codes
+from app.models import User
+
+
+@click.group(name="test-data", help="Generate and destroy test data")
+def test_data_group():
+    pass
+
+
+class test_data_command:
+    def __init__(self, name=None):
+        self.name = name
+
+    def __call__(self, func):
+        # we need to call the flask with_appcontext decorator to ensure the config is loaded, db connected etc etc.
+        # we also need to use functools.wraps to carry through the names and docstrings etc of the functions.
+        # Then we need to turn it into a click.Command - that's what command_group.add_command expects.
+        @click.command(name=self.name)
+        @functools.wraps(func)
+        @flask_cli.with_appcontext
+        def wrapper(*args, **kwargs):
+            return func(*args, **kwargs)
+
+        test_data_group.add_command(wrapper)
+
+        return wrapper
+
+
+def setup_test_data_commands(application):
+    application.cli.add_command(test_data_group)
+
+
+@test_data_command()
+@click.option(
+    "-u",
+    "--user_email_prefix",
+    required=True,
+    help="""
+    Functional test user email prefix. eg "notify-test-preview"
+""",
+)  # noqa
+def purge_functional_test_data(user_email_prefix):
+    """
+    Remove non-seeded functional test data
+
+    users, services, etc. Give an email prefix. Probably "notify-test-preview".
+    """
+    users = User.query.filter(User.email_address.like("{}%".format(user_email_prefix))).all()
+    for usr in users:
+        # Make sure the full email includes a uuid in it
+        # Just in case someone decides to use a similar email address.
+        try:
+            uuid.UUID(usr.email_address.split("@")[0].split("+")[1])
+        except ValueError:
+            print("Skipping {} as the user email doesn't contain a UUID.".format(usr.email_address))
+        else:
+            services = dao_fetch_all_services_by_user(usr.id)
+            if services:
+                for service in services:
+                    delete_service_and_all_associated_db_objects(service)
+            else:
+                delete_user_verify_codes(usr)
+                delete_model_user(usr)
